+++ runner/runner.go
@@ -1,16 +1,17 @@
-// Package runner executes one Filament ingestion run.
+// Package runner executes one Filament ingestion run. It is the single
+// implementation of the run sequence: resolve providers, open the sink, drive
+// the pipeline, then commit or abort, publishing lifecycle facts throughout.
+//
+// Both dispatch backends call RunOne — the engine module runs it in-process,
+// the worker binary runs it once per Job — so where a run executes is a
+// deployment choice and never a behavioral one. State transitions are not
+// written here; they are folded from the published facts by the tracker. The
+// sole exception is the resumable cursor seed in resolveExtractor.
 package runner
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
-	"runtime/debug"
-	"sort"
-	"strings"
-	"sync"
-	"sync/atomic"
 	"time"
 
 	"github.com/galaxy-io/filament"
@@ -19,7 +20,13 @@ import (
 	"github.com/galaxy-io/filament/pipeline"
 )
 
-// Deps are the process-local dependencies needed to execute one run.
+// abortWait bounds sink cleanup after a failed run. Detached from run
+// cancellation so shutdown still cleans staged data, bounded so a dead sink
+// cannot stall it indefinitely.
+const abortWait = 30 * time.Second
+
+// Deps are the process-local dependencies needed to execute one run. Tracer is
+// optional; a nil one skips span reporting.
 type Deps struct {
 	Bus       eventbus.Bus
 	DataStore filament.DataStore
@@ -27,6 +34,7 @@ type Deps struct {
 	Sources   filament.SourceRegistry
 	Sinks     filament.SinkRegistry
 	Log       filament.Logger
+	Tracer    filament.Tracer
 }
 
 // SpecFromState builds the RunSpec to execute from a persisted run state.
@@ -41,6 +49,24 @@ func SpecFromState(s filament.RunState) filament.RunSpec {
 	}
 }
 
+// ShouldRun reports whether a persisted run is this attempt's to execute. Only a
+// freshly requested or resumable-partial run qualifies, so a redelivered trigger
+// for a run already running or finished is a no-op rather than duplicated work.
+//
+// A completed CDC run is the exception: it is a catch-up cycle by construction,
+// so re-requesting it continues the change stream from its persisted cursor,
+// draining the source up to a fresh watermark and completing again.
+func ShouldRun(state filament.RunState) bool {
+	switch state.Status {
+	case filament.RunRequested, filament.RunPartial:
+		return true
+	case filament.RunCompleted:
+		return state.Request.IngestionType.OrDefault() == filament.IngestionCDC
+	default:
+		return false
+	}
+}
+
 // RunOne executes a single extraction end-to-end: resolve providers, open the
 // sink, drive the pipeline with the source, then commit or abort. It publishes
 // run.started first and exactly one terminal fact (run.completed | run.failed |
@@ -48,7 +74,18 @@ func SpecFromState(s filament.RunState) filament.RunSpec {
 //
 //nolint:funlen // the run lifecycle reads best as one sequence
 func RunOne(ctx context.Context, deps Deps, spec filament.RunSpec) {
+	var span filament.Span
+	if deps.Tracer != nil {
+		ctx, span = deps.Tracer.Start(ctx, "filament.run")
+		defer span.End()
+		span.SetAttr("run", string(spec.Run))
+		span.SetAttr("tenant", string(spec.Tenant))
+		span.SetAttr("source", spec.Source.Provider)
+		span.SetAttr("sink", spec.Sink.Provider)
+	}
+
 	em := newEmitter(ctx, deps.Bus, deps.Log, spec.Tenant, spec.Run)
+	em.span = span
 	emit(em, events.RunStarted, "", events.RunStartedEvent{})
 
 	if err := ResolveConfigRefs(ctx, deps.Secrets, &spec); err != nil {
@@ -93,6 +130,9 @@ func RunOne(ctx context.Context, deps Deps, spec filament.RunSpec) {
 		return
 	}
 
+	// A schema-aware sink needs typed DDL before any write. When the source can
+	// supply per-resource schemas, ensure each one up front so a Schematized sink
+	// (e.g. iceberg, postgres) creates/evolves its tables before extraction.
 	if err := ensureSchemas(ctx, src, snk, spec); err != nil {
 		if aerr := snk.Abort(ctx); aerr != nil && deps.Log != nil {
 			deps.Log.Error("runner: sink abort", aerr, filament.Field{Key: "run", Value: string(spec.Run)})
@@ -101,6 +141,7 @@ func RunOne(ctx context.Context, deps Deps, spec filament.RunSpec) {
 		return
 	}
 
+	// Announce the resources this run will touch (when known up front).
 	for _, res := range spec.Resources {
 		emit(em, events.ResourceStarted, res, events.ResourceStartedEvent{})
 	}
@@ -126,6 +167,9 @@ func RunOne(ctx context.Context, deps Deps, spec filament.RunSpec) {
 	})
 	p.Start(ctx)
 
+	// Extract on its own goroutine so the writer can apply backpressure through
+	// the inlet. CloseIngest after Extract returns drains the batcher; Wait then
+	// blocks until the writer finishes. A panicking source is contained here.
 	extractErrCh := make(chan error, 1)
 	go func() {
 		extractErrCh <- safeCall(func() error {
@@ -142,6 +186,9 @@ func RunOne(ctx context.Context, deps Deps, spec filament.RunSpec) {
 	waitErr := p.Wait()
 	extractErr := <-extractErrCh
 
+	// A write failure (waitErr) is the root cause and takes precedence: when the
+	// pipeline fails, the inlet returns it to Extract too, so extractErr is just
+	// the echo.
 	runErr := waitErr
 	if runErr == nil {
 		runErr = extractErr
@@ -149,21 +196,28 @@ func RunOne(ctx context.Context, deps Deps, spec filament.RunSpec) {
 
 	resources := resolveResources(spec.Resources, em.seenResources())
 
+	// The run is ending: detach fact publishing from run cancellation so the
+	// terminal facts survive host shutdown — the obituary must outlive the
+	// death, or the row strands in RunRunning forever.
+	defer em.finish()()
+
 	if runErr != nil {
-		if isResumableRun(plan) {
-			for _, res := range resources {
-				emit(em, events.ResourceFailed, res, events.ResourceFailedEvent{Error: runErr.Error()})
-			}
-			em.partial(runErr)
-			return
-		}
-		if err := snk.Abort(ctx); err != nil && deps.Log != nil {
-			deps.Log.Error("runner: sink abort", err, filament.Field{Key: "run", Value: string(spec.Run)})
-		}
 		for _, res := range resources {
 			emit(em, events.ResourceFailed, res, events.ResourceFailedEvent{Error: runErr.Error()})
 		}
+		if isResumableRun(plan) {
+			em.partial(runErr)
+			return
+		}
 		em.fail(runErr)
+		// Abort after the obituary, on its own detached context: cleanup must
+		// not eat the terminal publish window, and a slow sink must not strand
+		// the row in RunRunning.
+		abortCtx, cancel := context.WithTimeout(context.WithoutCancel(ctx), abortWait)
+		defer cancel()
+		if err := snk.Abort(abortCtx); err != nil && deps.Log != nil {
+			deps.Log.Error("runner: sink abort", err, filament.Field{Key: "run", Value: string(spec.Run)})
+		}
 		return
 	}
 
@@ -177,374 +231,4 @@ func RunOne(ctx context.Context, deps Deps, spec filament.RunSpec) {
 	}
 	records, bytes := em.runTotals()
 	emit(em, events.RunCompleted, "", events.RunCompletedEvent{Records: records, Bytes: bytes})
-}
-
-// ResolveConfigRefs resolves opaque config and field references into the
-// process-local RunSpec copy immediately before connector configuration.
-func ResolveConfigRefs(ctx context.Context, secrets filament.Secrets, spec *filament.RunSpec) error {
-	if err := resolveRefConfig(ctx, secrets, &spec.Source, spec.Tenant, "source"); err != nil {
-		return err
-	}
-	if err := resolveRefConfig(ctx, secrets, &spec.Sink, spec.Tenant, "sink"); err != nil {
-		return err
-	}
-	if err := resolveSecretRefs(ctx, secrets, &spec.Source, spec.Tenant, "source"); err != nil {
-		return err
-	}
-	if err := resolveSecretRefs(ctx, secrets, &spec.Sink, spec.Tenant, "sink"); err != nil {
-		return err
-	}
-	return nil
-}
-
-// resolveSecretRefs reads each of the ref's declared secrets and injects the
-// plaintext value into the provider config under the mapped field. Every ref is
-// tenant-scoped first, so a spec cannot read another tenant's connection secrets.
-func resolveSecretRefs(ctx context.Context, secrets filament.Secrets, ref *filament.Ref, tenant filament.TenantID, role string) error {
-	if len(ref.SecretRefs) == 0 {
-		return nil
-	}
-	if secrets == nil {
-		return fmt.Errorf("%s %q has secret refs but no secrets store is configured", role, ref.Provider)
-	}
-	if ref.Config == nil {
-		ref.Config = make(map[string]any, len(ref.SecretRefs))
-	}
-	for field, name := range ref.SecretRefs {
-		if err := filament.ValidateConnectionSecretRef(name, tenant); err != nil {
-			return fmt.Errorf("resolve %s secret for field %q: %w", role, field, err)
-		}
-		secret, err := secrets.Read(ctx, name)
-		if err != nil {
-			return fmt.Errorf("resolve %s secret %q for field %q: %w", role, name, field, err)
-		}
-		setConfigPath(ref.Config, field, string(secret.Value))
-	}
-	return nil
-}
-
-// setConfigPath assigns value at a dotted field path, rebuilding objects that
-// were removed when their only submitted values were secrets.
-func setConfigPath(cfg map[string]any, path string, value any) {
-	parts := strings.Split(path, ".")
-	current := cfg
-	for _, part := range parts[:len(parts)-1] {
-		nested, ok := current[part].(map[string]any)
-		if !ok {
-			nested = make(map[string]any)
-			current[part] = nested
-		}
-		current = nested
-	}
-	current[parts[len(parts)-1]] = value
-}
-
-func resolveRefConfig(ctx context.Context, secrets filament.Secrets, ref *filament.Ref, tenant filament.TenantID, role string) error {
-	if ref.ConfigRef == "" || len(ref.Config) > 0 {
-		return nil
-	}
-	if secrets == nil {
-		return fmt.Errorf("%s %q has config ref %q but no secrets store is configured", role, ref.Provider, ref.ConfigRef)
-	}
-	if err := filament.ValidateConnectionSecretRef(ref.ConfigRef, tenant); err != nil {
-		return fmt.Errorf("read %s config ref: %w", role, err)
-	}
-	secret, err := secrets.Read(ctx, ref.ConfigRef)
-	if err != nil {
-		return fmt.Errorf("read %s config ref %q: %w", role, ref.ConfigRef, err)
-	}
-	var cfg map[string]any
-	if err := json.Unmarshal(secret.Value, &cfg); err != nil {
-		return fmt.Errorf("decode %s config ref %q as JSON object: %w", role, ref.ConfigRef, err)
-	}
-	ref.Config = cfg
-	return nil
-}
-
-type extractorFunc func(context.Context, filament.RecordSink, filament.ExtractOpts) error
-
-func resolveExtractor(ctx context.Context, ds filament.DataStore, src filament.Source, spec filament.RunSpec, plan filament.IngestionPlan) (extractorFunc, error) {
-	if plan.Type == filament.IngestionCDC {
-		changes, ok := src.(filament.ChangeSource)
-		if !ok {
-			return nil, fmt.Errorf("source %q does not support CDC extraction", spec.Source.Provider)
-		}
-		checkpoints, err := loadChangeCheckpoints(ctx, ds, spec)
-		if err != nil {
-			return nil, err
-		}
-		return func(ctx context.Context, sink filament.RecordSink, opts filament.ExtractOpts) error {
-			return changes.ExtractChanges(ctx, sink, filament.ChangeExtractOpts{
-				Resources:   opts.Resources,
-				Checkpoints: checkpoints,
-				Limit:       opts.Limit,
-			})
-		}, nil
-	}
-	if isResumableRun(plan) {
-		resumable, ok := src.(filament.Resumable)
-		if !ok {
-			return nil, fmt.Errorf("source %q does not support resumable extraction", spec.Source.Provider)
-		}
-		incremental := plan.SourcePolicy.Mode == filament.ModeIncremental
-		prev := make(map[string]filament.Checkpoint, len(spec.Resources))
-		for _, resource := range spec.Resources {
-			var cp filament.Checkpoint
-			var err error
-			if incremental {
-				key, valid := spec.ResourceCheckpointKey(resource)
-				if !valid {
-					return nil, fmt.Errorf("incremental resource %q requires a versioned pipeline route", resource)
-				}
-				var state filament.ResourceCheckpointState
-				state, err = ds.LoadResourceCheckpoint(ctx, key)
-				cp = state.Checkpoint
-			} else {
-				cp, err = ds.LoadCheckpoint(ctx, spec.Run, resource)
-			}
-			if err == nil {
-				prev[resource] = cp
-			} else if err != nil && !errors.Is(err, filament.ErrNotFound) {
-				return nil, fmt.Errorf("load checkpoint %q: %w", resource, err)
-			}
-		}
-		var resumePlan map[string]filament.Checkpoint
-		var err error
-		if incremental {
-			planner, ok := src.(filament.IncrementalPlanner)
-			if !ok {
-				return nil, fmt.Errorf("source %q does not support incremental planning", spec.Source.Provider)
-			}
-			resumePlan, err = planner.PlanIncremental(ctx, spec.Resources, prev, spec.CursorConfigs)
-		} else {
-			planner, ok := src.(filament.ResumePlanner)
-			if !ok {
-				return nil, fmt.Errorf("source %q does not support resumable planning", spec.Source.Provider)
-			}
-			resumePlan, err = planner.PlanResume(ctx, spec.Resources, prev)
-		}
-		if err != nil {
-			return nil, fmt.Errorf("plan resume: %w", err)
-		}
-		for _, cp := range resumePlan {
-			if cp == nil {
-				continue
-			}
-			var saveErr error
-			if incremental {
-				key, _ := spec.ResourceCheckpointKey(cp.Resource())
-				saveErr = ds.SaveResourceCheckpoint(ctx, filament.ResourceCheckpointState{Key: key, Run: spec.Run, Checkpoint: cp})
-			} else {
-				saveErr = ds.SaveCheckpoint(ctx, spec.Run, cp)
-			}
-			if saveErr != nil {
-				return nil, fmt.Errorf("seed checkpoint %q: %w", cp.Resource(), saveErr)
-			}
-		}
-		return func(ctx context.Context, sink filament.RecordSink, opts filament.ExtractOpts) error {
-			return resumable.ExtractFrom(ctx, sink, opts, resumePlan)
-		}, nil
-	}
-	return func(ctx context.Context, sink filament.RecordSink, opts filament.ExtractOpts) error {
-		return src.Extract(ctx, sink, opts)
-	}, nil
-}
-
-func loadChangeCheckpoints(ctx context.Context, ds filament.DataStore, spec filament.RunSpec) (map[string]filament.Checkpoint, error) {
-	out := make(map[string]filament.Checkpoint, len(spec.Resources))
-	for _, resource := range spec.Resources {
-		var cp filament.Checkpoint
-		var err error
-		if key, ok := spec.ResourceCheckpointKey(resource); ok {
-			var state filament.ResourceCheckpointState
-			state, err = ds.LoadResourceCheckpoint(ctx, key)
-			cp = state.Checkpoint
-		} else {
-			cp, err = ds.LoadCheckpoint(ctx, spec.Run, resource)
-		}
-		if err == nil {
-			out[resource] = cp
-			continue
-		}
-		if !errors.Is(err, filament.ErrNotFound) {
-			return nil, fmt.Errorf("load checkpoint %q: %w", resource, err)
-		}
-	}
-	if len(out) == 0 {
-		return nil, nil
-	}
-	return out, nil
-}
-
-func isResumableRun(plan filament.IngestionPlan) bool {
-	return plan.SourcePolicy.Checkpointing != filament.CheckpointNone && !plan.RequiresCDC
-}
-
-func safeCall(fn func() error) (err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			err = fmt.Errorf("source panicked: %v\n%s", r, debug.Stack())
-		}
-	}()
-	return fn()
-}
-
-func ensureSchemas(ctx context.Context, src filament.Source, snk filament.Sink, spec filament.RunSpec) error {
-	sch, ok := snk.(filament.Schematized)
-	if !ok {
-		return nil
-	}
-	prov, ok := src.(filament.SchemaProvider)
-	if !ok {
-		return fmt.Errorf("sink %q requires a schema but source %q provides none", spec.Sink.Provider, spec.Source.Provider)
-	}
-	for _, res := range spec.Resources {
-		schema, err := prov.Schema(ctx, res)
-		if err != nil {
-			return fmt.Errorf("schema for %q: %w", res, err)
-		}
-		if err := sch.EnsureSchema(ctx, res, schema); err != nil {
-			return fmt.Errorf("ensure schema for %q: %w", res, err)
-		}
-	}
-	return nil
-}
-
-type emitter struct {
-	ctx    context.Context
-	bus    eventbus.Bus
-	log    filament.Logger
-	tenant filament.TenantID
-	run    filament.RunID
-
-	seq atomic.Uint64
-
-	mu         sync.Mutex
-	runRecords int64
-	runBytes   int64
-	res        map[string]*tally
-}
-
-type tally struct {
-	records int64
-	bytes   int64
-}
-
-func newEmitter(ctx context.Context, bus eventbus.Bus, log filament.Logger, tenant filament.TenantID, run filament.RunID) *emitter {
-	return &emitter{ctx: ctx, bus: bus, log: log, tenant: tenant, run: run, res: map[string]*tally{}}
-}
-
-func (e *emitter) next() uint64 { return e.seq.Add(1) }
-
-func (e *emitter) publish(f events.Fact) {
-	if d, ok := f.Data.(events.BatchWrittenEvent); ok {
-		e.mu.Lock()
-		e.runRecords += d.Records
-		e.runBytes += d.Bytes
-		t := e.res[f.Resource]
-		if t == nil {
-			t = &tally{}
-			e.res[f.Resource] = t
-		}
-		t.records += d.Records
-		t.bytes += d.Bytes
-		e.mu.Unlock()
-	}
-	if e.log != nil {
-		records, bytes, errMsg := factProgress(f)
-		fields := []filament.Field{
-			{Key: "run", Value: string(f.Run)},
-			{Key: "resource", Value: f.Resource},
-			{Key: "records", Value: records},
-			{Key: "bytes", Value: bytes},
-		}
-		if errMsg != "" {
-			fields = append(fields, filament.Field{Key: "error", Value: errMsg})
-		}
-		e.log.Info(f.Name, fields...)
-	}
-	if err := events.Publish(e.ctx, e.bus, f); err != nil && e.log != nil {
-		e.log.Error("runner: publish fact", err, filament.Field{Key: "type", Value: f.Name})
-	}
-}
-
-// factProgress flattens a typed payload's progress counters and error for logging.
-func factProgress(f events.Fact) (records, bytes int64, errMsg string) {
-	switch d := f.Data.(type) {
-	case events.BatchBufferedEvent:
-		return d.Records, d.Bytes, ""
-	case events.BatchWrittenEvent:
-		return d.Records, d.Bytes, ""
-	case events.ResourceCompletedEvent:
-		return d.Records, d.Bytes, ""
-	case events.RunCompletedEvent:
-		return d.Records, d.Bytes, ""
-	case events.ResourceFailedEvent:
-		return 0, 0, d.Error
-	case events.RunFailedEvent:
-		return 0, 0, d.Error
-	case events.RunPartialEvent:
-		return 0, 0, d.Error
-	default:
-		return 0, 0, ""
-	}
-}
-
-// emit stamps and publishes a runner-originated fact for a run or resource.
-// (A free function: Go methods cannot take type parameters.)
-func emit[T any](e *emitter, t events.EventType[T], resource string, data T) {
-	e.publish(events.NewFact(t, events.Envelope{
-		Tenant:   e.tenant,
-		Run:      e.run,
-		Resource: resource,
-		Seq:      e.next(),
-		At:       time.Now(),
-	}, data))
-}
-
-func (e *emitter) fail(err error) {
-	if e.log != nil {
-		e.log.Error("runner: run failed", err, filament.Field{Key: "run", Value: string(e.run)})
-	}
-	emit(e, events.RunFailed, "", events.RunFailedEvent{Error: err.Error()})
-}
-
-func (e *emitter) partial(err error) {
-	if e.log != nil {
-		e.log.Error("runner: run partial", err, filament.Field{Key: "run", Value: string(e.run)})
-	}
-	emit(e, events.RunPartial, "", events.RunPartialEvent{Error: err.Error()})
-}
-
-func (e *emitter) resourceTally(name string) (records, bytes int64) {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-	if t := e.res[name]; t != nil {
-		return t.records, t.bytes
-	}
-	return 0, 0
-}
-
-func (e *emitter) runTotals() (records, bytes int64) {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-	return e.runRecords, e.runBytes
-}
-
-func (e *emitter) seenResources() []string {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-	out := make([]string, 0, len(e.res))
-	for name := range e.res {
-		out = append(out, name)
-	}
-	sort.Strings(out)
-	return out
-}
-
-func resolveResources(requested, seen []string) []string {
-	if len(requested) > 0 {
-		return requested
-	}
-	return seen
 }
